value groups
I realised the previous implementation could still use the GroupEnd type and have the same user-facing API.
The only thing you'd lose is the ability to skip a group via manipulation of the index; In the GroupEnd implementation, to reach the end, you must go to the end.

I don't think the ability to skip really matters here.

This is more simple, and uses less memory.
I suppose it is less familiar than what I would expect (what I would expect is something like the proc-macro api), but this should be just as ergonomic in practice.