no escape sequences
I've decided to drop support for escape sequences for now. I was implementing them because you would expect to be able to use them, and I still think that they should be implemented "eventually".

However, for my purposes, I wouldn't use them.

Also the ANSI escape sequences seem to be more comprehensive than what I was looking for. They seem to allow for manipulating how a terminal would print something.

https://crates.io/crates/unicode_escape
This is more what I was looking for, so If I come back to this I'll probably use that.
(I do not want to implement my own escapes).