escapes skeleton
I decided to implement escapes.

What made it unappealing at first was that I couldn't find the bounds of the task; I looked for a library to handle it for me, because then I could off-load the responsibility of implementing whatever the "standard" was to that library.
None of them were suitable though.

<https://www.json.org/json-en.html>
Eventually I found this. Official JSON documentation; it details how escapes are supposed to work in JSON.

Doing it as JSON does it keeps it consistent with the rest of the locatch tools.

Why JSON?
Afterall, there are modern alternatives (JSON-5, RON, and probably many more).
I am not necessarily against those, but JSON is plain and common, and just implementing what is plain and common is good enough for now.

Locatch is intended to be "ad-hoc" and the plain-ness of JSON fits with that; It keeps the implementation and complexity low.
So following standards set by JSON, makes designing this much easier.

I tackle escapes now, because it provides relief from the question of how they should be implemented; There is security in finding an answer now, even if the feature isn't really needed by me for the foreseeable future.