compiler issue
The borrow checker doesn't like this.
Tedious, theoretically we are okay, but we have to find a way to make the compiler accept it.

I believe it does not accept it, because it is not dropping the mutable borrow, due to the output having a lifetime shared with the iterator?
The reason this is theoretically okay, is because the iterator does not contain a mutable reference to the slice; It does not mutate the slice.

I think the fix is to introduce another lifetime for the output.