not tests passing (fin)
Might want some assertion utilities or something, because that is a whole lot of code in the tests there.
Was okay to write it, but I wouldn't want to read it.

Also, there's still technically more tests to make.
Errors, and then making tests to test that it errors as expected.
Erroring is a part of the design, because the from function validates the integrity of the data so that the reader doesn't have to check it itself.

That said, this is good enough for my purposes, for now, so I will probably be moving on.
(My purposes don't need this thing to be highly usable, because I already understand it well)